addCourse, addStudent error잡기완료

+++ Grpc-Data/src/main/java/com/example/grpc/data/CrudMethods.java
@@ -1,37 +1,68 @@
 package com.example.grpc.data;
 
-import com.example.exception.AlreadyExistIdException;
+import com.example.exception.*;
 import com.example.grpc.*;
 import io.grpc.stub.StreamObserver;
+import lombok.extern.slf4j.Slf4j;
 
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.nio.file.StandardOpenOption;
 import java.util.Map;
+import java.util.logging.Logger;
 
+@Slf4j
 public class CrudMethods {
-    public void writeStudent(StudentInfoString studentInfoString) throws AlreadyExistIdException {
+
+    public void writeStudent(StudentInfoString studentInfoString) throws AlreadyExistStudentIDException, NotExistCourseIDException {
         String studentInfo = studentInfoString.getStudentInfo();
         String studentID = studentInfo.split(" ")[0];
         studentInfo = "\n\n"+studentInfoString.getStudentInfo();
-        match(studentID, SCRSProperties.STUDENT_LIST_PATH);
-
+        if(match(studentID, SCRSProperties.STUDENT_LIST_PATH)) throw new AlreadyExistStudentIDException();
         try {
             Files.write(Paths.get(SCRSProperties.STUDENT_LIST_PATH), studentInfo.getBytes(), StandardOpenOption.APPEND);
+        } catch (IOException e) {
+            log.error("IOException");
+        }
+    }
+
+    public void writeCourse(CourseInfoString courseInfoString) throws AlreadyExistCourseIDException, NotExistCourseIDException {
+        String courseInfo = courseInfoString.getCourseInfo();
+        String[] courseInfoArray = courseInfo.split(" ");
+        String courseID = courseInfoArray[0];
+        if(match(courseID, SCRSProperties.COURSE_LIST_PATH)) throw new AlreadyExistCourseIDException();
+        if(courseInfoArray.length>3)
+            for(int i=courseInfoArray.length-1; i>=3; i--)
+                if(!match(courseInfoArray[i], SCRSProperties.COURSE_LIST_PATH)) throw new NotExistCourseIDException();
+
+        courseInfo = "\n\n"+courseInfo;
+        try {
+            Files.write(Paths.get(SCRSProperties.COURSE_LIST_PATH), courseInfo.getBytes(), StandardOpenOption.APPEND);
         } catch (IOException e) {
             e.printStackTrace();
         }
     }
 
-    public boolean delete(String id, String fileName){
+    public boolean delete(String id, String fileName, StreamObserver<Message> responseObserver){
+        try{
+            if(!match(id, fileName)){
+                throw new NotExistIDException();
+            }
+        }catch(NotExistIDException e){
+            Message msg = Message.newBuilder().setMsg("NOTexistID").build();
+            responseObserver.onNext(msg);
+            responseObserver.onCompleted();
+            e.printStackTrace();
+            return false;
+        }
 
         BufferedReader br = null;
         String all = "";
         try {
             br = new BufferedReader(new FileReader(fileName));
         } catch (FileNotFoundException e) {
-            e.printStackTrace();
+            log.error("FileNotFoundException");
             return false;
         }
         try{
@@ -70,29 +101,18 @@ public class CrudMethods {
         return true;
     }
 
-    public void writeCourse(CourseInfoString courseInfoString) throws AlreadyExistIdException {
-        String courseInfo = courseInfoString.getCourseInfo();
-        String[] courseInfoArray = courseInfo.split(" ");
-        String courseID = courseInfoArray[0];
-        courseInfo = "\n\n"+courseInfo;
-        match(courseID, SCRSProperties.COURSE_LIST_PATH);
 
-        try {
-            Files.write(Paths.get(SCRSProperties.COURSE_LIST_PATH), courseInfo.getBytes(), StandardOpenOption.APPEND);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
 
     public boolean update(String studentId, Map<Integer, String> courseIDListMap, StreamObserver<Message> responseObserver) {
         try {/*사용자가 입력한 studentId가 유효한지 확인*/
-            if(!match(studentId, SCRSProperties.STUDENT_LIST_PATH, responseObserver)){
-                Message msg = Message.newBuilder().setMsg("NOTexistIDstd").build();
-                responseObserver.onNext(msg);
-                responseObserver.onCompleted();
+            if(!match(studentId, SCRSProperties.STUDENT_LIST_PATH)){
+                throw new NotExistStudentIDException();
             }
-        } catch (AlreadyExistIdException e) {
-            e.printStackTrace();
+        } catch (NotExistStudentIDException e) {
+            Message msg = Message.newBuilder().setMsg("NOTexistIDstd").build();
+            responseObserver.onNext(msg);
+            responseObserver.onCompleted();
+            log.error("사용자가 존재하지 않는 학생번호를 입력했습니다.");
             return false;
         }
 
@@ -103,23 +123,14 @@ public class CrudMethods {
             while (br.ready()) {
                 String line = br.readLine();
                 String[] splitLine = line.split(" ");
-                /*수상신청을 해줘야하는 학생 발견!*/
                 if(splitLine[0].equals(studentId)){
                     System.out.println(line);
-                    /*이미 수강한 강좌는 아닌지 확인*/
                     boolean notTake = checkAlreadyTake(courseIDListMap, splitLine, responseObserver);
-                    System.out.println("notTake:"+notTake);
-                    /*수강신청한 courseID가 실제로 존재하는 courseID인지 확인하기*/
                     boolean exist = isExistCourse(courseIDListMap, responseObserver);
-                    System.out.println("exist:"+exist);
-                    if(notTake&&exist){
-                        all += addCourseId(courseIDListMap, line)+ "\n";
-                    }else{
-                        all += line;
-                    }
-                }else{
-                    all += line + "\n";
-                }
+                    if(notTake&&exist) all += addCourseId(courseIDListMap, line)+ "\n";
+                    else all += line;
+                }else all += line + "\n";
+
             }
         }
         catch(IOException e){
@@ -156,36 +167,43 @@ public class CrudMethods {
         for (Integer i : courseIDListMap.keySet()) {
             String courseId = courseIDListMap.get(i);
             try {
-                if(!match(courseId, SCRSProperties.COURSE_LIST_PATH, responseObserver)){
-                    Message msg = Message.newBuilder().setMsg("NOTexistIDcourse").build();
-                    responseObserver.onNext(msg);
-                    responseObserver.onCompleted();
-                    return false;
+                if(!match(courseId, SCRSProperties.COURSE_LIST_PATH)){
+                    throw new NotExistCourseIDException();
                 }
-            } catch (AlreadyExistIdException e) {
-                e.printStackTrace();
+            } catch (NotExistCourseIDException e) {
+                Message msg = Message.newBuilder().setMsg("NOTexistIDcourse").build();
+                responseObserver.onNext(msg);
+                responseObserver.onCompleted();
+                log.error("사용자가 존재하지 않는 강좌번호를 입력했습니다.");
+                return false;
             }
         }
         return true;
     }
 
     public boolean checkAlreadyTake(Map<Integer, String> courseIDListMap, String[] splitLine, StreamObserver<Message> responseObserver) {
-        for (Integer i : courseIDListMap.keySet()) {
-            String courseId = courseIDListMap.get(i);
-            for(int j=splitLine.length;j>2; j--){
-                if(splitLine.equals(courseId)){
-                    Message msg = Message.newBuilder().setMsg("alreadyEcourse").build();
-                    responseObserver.onNext(msg);
-                    responseObserver.onCompleted();
-                    return false;
+        try {
+            for (Integer i : courseIDListMap.keySet()) {
+                String courseId = courseIDListMap.get(i);
+                for (int j = splitLine.length; j > 2; j--) {
+                    if (splitLine.equals(courseId)) {
+                        throw new AlreadyTakeCourseIDException();
+                    }
                 }
             }
+        }catch(AlreadyTakeCourseIDException e){
+            Message msg = Message.newBuilder().setMsg("alreadyEcourse").build();
+            responseObserver.onNext(msg);
+            responseObserver.onCompleted();
+            log.error("사용자가 이미 수강한 강좌의 강좌번호로 수강신청을 시도했습니다.");
+            return false;
         }
+
         return true;
     }
 
-    //새로 입력한 info의 id와 이미 존재하는 fileName의 id를 비교하고 이미 있는 id이면 AlreadyExistIdException을 던잔다.
-    public boolean match(String id, String fileName, StreamObserver<Message> responseObserver) throws AlreadyExistIdException{
+    //id와 일치하면 true반환
+    public boolean match(String id, String fileName){
         BufferedReader objStudentFile = null;
         try {
             objStudentFile = new BufferedReader(new FileReader(fileName));
@@ -208,34 +226,6 @@ public class CrudMethods {
         return false;
     }
 
-    public boolean match(String id, String fileName) throws AlreadyExistIdException{
-        int i = 0;
-        BufferedReader objStudentFile = null;
-        try {
-            objStudentFile = new BufferedReader(new FileReader(fileName));
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        }
-        try{
-            while (objStudentFile.ready()) {
-                assert objStudentFile != null;
-                String line = objStudentFile.readLine();
-                String[] alreadyId = line.split(" ");
-                if(alreadyId[0].equals(id)){
-                    return false;
-                }
-            }
-        }
-        catch(IOException e){
-            e.printStackTrace();
-        }
-        try {
-            objStudentFile.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return true;
-    }
 
     public DataResponse.Builder putStudentInfo(DataResponse.Builder builder){
         int i = 0;
